Conflicts:
	src/main/java/project/web/mvc/controller/MyPageController.java
	src/main/java/project/web/mvc/repository/WishListRepository.java
	src/main/java/project/web/mvc/service/CartService.java
	src/main/resources/application.properties

	wishlist

+++ src/main/java/project/web/mvc/repository/NoticeRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import project.web.mvc.domain.Notice;
 
-public interface NoticeRepository extends CrudRepository<Notice, Integer> {
+public interface NoticeRepository extends CrudRepository<Notice, Long> {
 
 }
